requirements.txt
ca-certificates=2023.12.12=haa95532_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
docopt=0.6.2=pypi_0
idna=3.6=pypi_0
mysql-connector-python=8.2.0=pypi_0
numpy=1.24.4=pypi_0
opencv-contrib-python=4.1.2.30=pypi_0
opencv-python=4.3.0.36=pypi_0
openssl=1.1.1w=h2bbff1b_0
pillow=8.1.0=pypi_0
pip=23.3.1=py38haa95532_0
pipreqs=0.4.13=pypi_0
protobuf=4.21.12=pypi_0
python=3.8.5=h5fd99cc_1
requests=2.31.0=pypi_0
setuptools=68.2.2=py38haa95532_0
sqlite=3.41.2=h2bbff1b_0
tzdata=2023c=h04d1e81_0
urllib3=2.1.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py38haa95532_0
yarg=0.1.9=pypi_0
python==3.8.5